backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Jose Pujol <josepujol21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
llama-index = "^0.10.20"
llama-index-packs-raptor = "^0.1.3"
llama-index-vector-stores-qdrant = "^0.1.4"
nest-asyncio = "^1.6.0"
python-dotenv = "^1.0.1"
chromadb = "^0.4.24"
llama-index-vector-stores-chroma = "^0.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
